# Codechange: Use FindVehiclesWithOrder when removing a road stop.