@LordAro 2024-02-20
What happens when the number of orders is significantly larger than the list of vehicles? :)