@PeterN 2024-02-27
Yup, 5000 road vehicles with 255 unshared orders each, is considerably slower this way.

Master: 80µs
This PR: 3,675µs